+++ CHANGELOG.md
@@ -4,8 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.1] - 2025-05-12
+## [1.3.0] - 2025-05-20
 
+### Added
+- The checkOptions method has been added to validate data in a set of data passed through an array.
+- Added checkRange to validate a value within a specific range.
+- Documentation has been created and hosted on gitbook.
+- New error messages have been added for the new methods (checkRange and checkOptions).
+
+### Changed
+- Updated README with new examples and information about the added types and documentation.
+
+## [1.2.1] - 2025-05-12
 ### Improved
 - Error messages now correctly identify the actual type of the received value.  
   **Example:**